AI

Description

Documentation

data Hiperparameters

Type representing MCTS algorithm hiperparameters

Constructors

Random

MCTS

Fields

balance :: Float

UCB c balance parameter

nIterations :: Int

number of MCTS iterations

type Generator = StdGen

Random number generator state used by the AI algorithm.

newGenerator :: Int -> Generator

For a given seed returns new generator.

getNextMove

Arguments

:: GameState
Game state to which AI player should respond

-> Hiperparameters
MCTS algorithm hiperparameters

-> Generator
Random number generator state

-> (Maybe Move, Generator)
(If there is a move possible then Just Move, if there is no move possible Nothing, State of the random number generator after calculating the move.)

Returns next move of the AI player.