Game

Description

Declarative representation of generalized Connect Four game for two players. Can represent any game as series of applications of appropiate functions. Game can be parametrized with number of rows, columns, and line length needed to win. Exported functions allow to create game, play it (check available moves and play moves) and check who wins.

Documentation

data BoardSetting

Parameters of the game.

Constructors

BoardSetting

Fields

nRows :: Int

Number of rows In classic Connect Four equal to 6

nColumns :: Int

Number of columns. In classic Connect Four equal to 7

lineToWin :: Int

Number of tokens placed consecutively either in horizontal, vertical, diagonal or antidiagonal line needed to win In classic Connect Four equal to 4

data Player

Type representing player Either Red or Blue

Constructors

Red

Blue

type BoardState = Seq (Maybe Player)

State of the board of the game. Sequence of elements - each of them represents an empty field (Nothing) or a field occupied by a player's token 'Just Player'. It is possible for it to store invalid states. For game representation to be correct, (length BoardState) has to be equal m * n, where m = number of rows, n = number of columns. Position (i, j) on the board is stored in the (m * i + j)th element of Sequence.

data GameState

State of the game. It is possible for it to store invalid states.

Constructors

GameState

Fields

boardSetting :: BoardSetting

Parameters of the game

boardState :: BoardState

State of the board

player :: Player

Which player turn it is (who will do next move)

type Move = Int

Move in the game. Value i means throwing token into i'th column. Columns are numbered from 0.

newGame :: BoardSetting -> GameState

Creates new game with empty board for provided parameters. Red player always has first move.

doesPlayerWin :: GameState -> Player -> Bool

Returns whether given player has a winning line O((2(m-k)(n-k) + m(n-k) + n(m-k)) * k * log(m*n))

whoWins :: GameState -> Maybe Player

Returns one of the winners of a given game. Nothing if there is no winnner yet 'Just Player' - one of the players having winning position For performance purposes doesn't check if game state is correct. If game wasn't played correctly, there can be more than one winner but always only one of them is returned. O(2 * (2(m-k)(n-k) + m(n-k) + n(m-k)) * k * log(m*n))

availableMoves :: GameState -> [Move]

Returns list of available moves for a given game. Available moves are columns that have their top field empty. For performance purposes doesn't check correctness of the game. O(n * log(m + n))

applyMove

Arguments

:: GameState
A game state. Correctness is not checked for performance reasons.

-> Move
Number of column in which the token will be placed.

-> Maybe GameState
If chosen column has at least one free place, then 'Just GameState' - game state with the move played. Token will be placed on the lowest free place int the column. Token's color will corespond to the player whose turn it was. If chosen column was full, or there was not enough columns, then Nothing.

For a given game state and a move plays that move and returns resulting game state if move was possible. O (m + n * log(m))

oponent :: Player -> Player

For given player a player that will move next.