UI

Description

Provides types and functions needed to render λeonard app with gi-gtk-declarative-app-simple framework.

Documentation

data State

Application states. Each state corresponds to one type of window application can render and values that have to be stored in that state.

Constructors

SettingsWindow
Main menu window

Fields

boardSetting :: BoardSetting

isRedComputer :: Bool

redDifficulty :: Difficulty

redSeed :: Int

isBlueComputer :: Bool

blueDifficulty :: Difficulty

blueSeed :: Int

speed :: Int

UserVSUserWindow
User VS User game window. Shows: game board, buttons with available moves, winner information when a player wins, "Back to menu" button. Users can play moves clicking corresponding buttons. After every move game board is immediately refreshed.

Fields

gameState :: GameState

UserVSComputerWindow
User VS User game window. Shows: game board, buttons with available moves when its Users turn, winner information when a player wins, "Back to menu" button. User can play moves clicking corresponding buttons. After every move game board is immediately refreshed, then immediately Computer plays his move and board is refreshed again.

Fields

gameState :: GameState

computerColor :: Player

computerSetting :: AISetting

ComputerVSComputerWindow
Computer VS Computer game window. Also called Simulation. Shows: game board, Pause/Resume button, winner information when a player wins, "Back to menu" button. Computer players play moves alternately. There is added a fixed delay speed between the moves. Immediately after every move the board is refreshed. Simulation can be paused and resumed any time.

Fields

gameState :: GameState

red :: AISetting

blue :: AISetting

speed :: Int

paused :: Bool

ErrorWindow
Error window containing single error message. State should not be reachable in the application.

Fields

message :: Text

data Event

Possible events. Each event depending on current state changes state and then does some IO action and finally maybe invokes other events.

Constructors

Closed

Error Text

MChanged Int
Settings modification events: (Evoked every time value is changed in the SettingsWindow) Each *Changed event is parametrized by the new value setting should take

NChanged Int

KChanged Int

RedTypeChanged Bool

BlueTypeChanged Bool

RedDifficultyChanged Difficulty

BlueDifficultyChanged Difficulty

RedSeedChanged Int

BlueSeedChanged Int

SpeedChanged Int

StartGame

EndGame
Game events: Should appear only in one of the game windows

MovePlayed Move

SimulationPaused

SimulationResumed

view' :: State -> AppView Window Event

Returns a render of a given state

update' :: State -> Event -> Transition State Event

For a given state and an event returns Transition to a next state. Transition contains next state and does some IO action that may result in a next event that will be evoken.

initialState' :: State

Initial state of the application. Application starts with the SettingsWindow.